upgrade test to actually run the test function
previous version of test program only served as a compile-time check;
it did not actually run the test function during 'make check'.
(Tested by temporarily adding ck_abort, which did not cause a failure
since the function was not getting run).

The tutorial for Check library is a bit unclear on this point, but
in order to make things labeled with START_TEST() actually be run,
it is necessary to complete the next tutorial stage that adds non-trivial
main() function.

"To run unit tests with Check, we must create some test cases, aggregate
them into a suite, and run them with a suite runner."
https://libcheck.github.io/check/doc/check_html/check_3.html#Creating-a-Suite