rename config.h to cantools_config.h and install it
When building the libraries and tools in-tree, config.h was
available so HAVE_CONFIG_H was defined and included,
which provided a bunch of other HAVE_* definitions so
the code could adapt to the current build platform.

However, since config.h was not being installed along with
the library headers, external callers of the library headers
could run into trouble because HAVE_* definitions
(e.g. HAVE_STDINT_H) were missing.

Address this by renaming config.h to cantools_config.h so
that it is unique enough to be installed with the other
headers. Then since always available, it can then be
unconditionally included by other headers so that external
callers won't get into trouble.